+++ FrameWork/zlib/Inflate.cs
@@ -377,7 +377,7 @@ namespace FrameWork
 			return Z_OK;
 		}
 		
-		private static byte[] mark = new byte[]{(byte) 0, (byte) 0, (byte) SupportClass.Identity(0xff), (byte) SupportClass.Identity(0xff)};
+		private static byte[] mark = new byte[]{0, 0, (byte) SupportClass.Identity(0xff), (byte) SupportClass.Identity(0xff)};
 		
 		internal int inflateSync(ZStream z)
 		{
